refactor code. Create datesManager
